Cast drawLine input arguments to int (#264)
* Cast drawLine input arguments to int

+++ qt_dotgraph/src/qt_dotgraph/dot_shapes.py
@@ -19,27 +19,28 @@ class QGraphicsBox3dItem(QAbstractGraphicsShapeItem):
                            self._bounding_box.height() - self._bounding_box.height() * 0.1)
         painter.drawRect(rectangle)
         # Top line
-        painter.drawLine(rectangle.topLeft().x() + self._bounding_box.height() * 0.1,
-                         self._bounding_box.topLeft().y(),
-                         self._bounding_box.topRight().x(),
-                         self._bounding_box.topRight().y())
+        painter.drawLine(int(rectangle.topLeft().x() + self._bounding_box.height() * 0.1),
+                         int(self._bounding_box.topLeft().y()),
+                         int(self._bounding_box.topRight().x()),
+                         int(self._bounding_box.topRight().y()))
         # Top left corner
-        painter.drawLine(rectangle.topLeft().x() + self._bounding_box.height() * 0.1,
-                         self._bounding_box.topLeft().y(),
-                         self._bounding_box.topLeft().x() + 1,
-                         rectangle.topLeft().y())
+        painter.drawLine(int(rectangle.topLeft().x() + self._bounding_box.height() * 0.1),
+                         int(self._bounding_box.topLeft().y()),
+                         int(self._bounding_box.topLeft().x() + 1),
+                         int(rectangle.topLeft().y()))
         # Top right corner
-        painter.drawLine(self._bounding_box.topRight().x(),
-                         self._bounding_box.topRight().y(),
-                         rectangle.topRight().x(),
-                         rectangle.topRight().y())
+        painter.drawLine(int(self._bounding_box.topRight().x()),
+                         int(self._bounding_box.topRight().y()),
+                         int(rectangle.topRight().x()),
+                         int(rectangle.topRight().y()))
         # Bottom right corner
-        painter.drawLine(rectangle.bottomRight().x() + 1,
-                         rectangle.bottomRight().y() - 1,
-                         self._bounding_box.bottomRight().x(),
-                         rectangle.bottomRight().y() - self._bounding_box.height() * 0.1)
+        painter.drawLine(int(rectangle.bottomRight().x() + 1),
+                         int(rectangle.bottomRight().y() - 1),
+                         int(self._bounding_box.bottomRight().x()),
+                         int(rectangle.bottomRight().y() - self._bounding_box.height() * 0.1))
         # Right line
-        painter.drawLine(self._bounding_box.topRight().x(),
-                         self._bounding_box.topRight().y(),
-                         self._bounding_box.topRight().x(),
-                         self._bounding_box.bottomRight().y() - self._bounding_box.height() * 0.1)
+        painter.drawLine(int(self._bounding_box.topRight().x()),
+                         int(self._bounding_box.topRight().y()),
+                         int(self._bounding_box.topRight().x()),
+                         int(self._bounding_box.bottomRight().y()
+                             - self._bounding_box.height() * 0.1))
